Move python module to python/tensorflow_cloud/core
