+++ python/tensorflow_cloud/core/containerize.py
@@ -144,7 +144,7 @@ class ContainerBuilder(object):
 
             # Add python 3 tag for TF version <= 2.1.0
             # https://hub.docker.com/r/tensorflow/tensorflow
-            if VERSION != 'latest':
+            if VERSION != "latest":
                 v = VERSION.split(".")
                 if float(v[0] + "." + v[1]) <= 2.1:
                     self.docker_base_image += "-py3"
