+++ python/tensorflow_cloud/core/gcp.py
