+++ python/tensorflow_cloud/core/machine_config.py
