+++ python/tensorflow_cloud/tests/integration/call_run_within_script_with_keras_fit_test.py
