+++ python/tensorflow_cloud/tests/tensorflow_cloud/validate_test.py
@@ -16,16 +16,16 @@
 import pytest
 import unittest
 
-from tensorflow_cloud import machine_config
-from tensorflow_cloud import validate
+from tensorflow_cloud.core import machine_config
+from tensorflow_cloud.core import validate
 
 
 class TestValidate(unittest.TestCase):
     def test_valid_args(self):
         validate.validate(
-            entry_point="tests/testdata/mnist_example_using_fit.py",
+            entry_point="python/tensorflow_cloud/tests/testdata/mnist_example_using_fit.py",
             distribution_strategy="auto",
-            requirements_txt="tests/testdata/requirements.txt",
+            requirements_txt="python/tensorflow_cloud/tests/testdata/requirements.txt",
             chief_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
             worker_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
             worker_count=1,
@@ -37,9 +37,9 @@ class TestValidate(unittest.TestCase):
         )
 
         validate.validate(
-            entry_point="tests/testdata/mnist_example_using_fit.py",
+            entry_point="python/tensorflow_cloud/tests/testdata/mnist_example_using_fit.py",
             distribution_strategy=None,
-            requirements_txt="tests/testdata/requirements.txt",
+            requirements_txt="python/tensorflow_cloud/tests/testdata/requirements.txt",
             chief_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
             worker_config=None,
             worker_count=0,
@@ -51,9 +51,9 @@ class TestValidate(unittest.TestCase):
         )
 
         validate.validate(
-            entry_point="tests/testdata/mnist_example_using_fit.ipynb",
+            entry_point="python/tensorflow_cloud/tests/testdata/mnist_example_using_fit.ipynb",
             distribution_strategy=None,
-            requirements_txt="tests/testdata/requirements.txt",
+            requirements_txt="python/tensorflow_cloud/tests/testdata/requirements.txt",
             chief_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
             worker_config=None,
             worker_count=0,
@@ -67,7 +67,7 @@ class TestValidate(unittest.TestCase):
         validate.validate(
             entry_point=None,
             distribution_strategy=None,
-            requirements_txt="tests/testdata/requirements.txt",
+            requirements_txt="python/tensorflow_cloud/tests/testdata/requirements.txt",
             chief_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
             worker_config=None,
             worker_count=0,
@@ -81,7 +81,7 @@ class TestValidate(unittest.TestCase):
         validate.validate(
             entry_point=None,
             distribution_strategy=None,
-            requirements_txt="tests/testdata/requirements.txt",
+            requirements_txt="python/tensorflow_cloud/tests/testdata/requirements.txt",
             chief_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
             worker_config=None,
             worker_count=0,
@@ -98,7 +98,7 @@ class TestValidate(unittest.TestCase):
             validate.validate(
                 entry_point="/mnist_example_using_fit.py",
                 distribution_strategy="auto",
-                requirements_txt="tests/testdata/requirements.txt",
+                requirements_txt="python/tensorflow_cloud/tests/testdata/requirements.txt",
                 chief_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
                 worker_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
                 worker_count=1,
@@ -113,7 +113,7 @@ class TestValidate(unittest.TestCase):
             validate.validate(
                 entry_point="/mnist_example_using_fit.txt",
                 distribution_strategy="auto",
-                requirements_txt="tests/testdata/requirements.txt",
+                requirements_txt="python/tensorflow_cloud/tests/testdata/requirements.txt",
                 chief_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
                 worker_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
                 worker_count=1,
@@ -127,7 +127,7 @@ class TestValidate(unittest.TestCase):
     def test_invalid_requirements_txt(self):
         with pytest.raises(ValueError, match=r"Invalid `requirements_txt`"):
             validate.validate(
-                entry_point="tests/testdata/mnist_example_using_fit.py",
+                entry_point="python/tensorflow_cloud/tests/testdata/mnist_example_using_fit.py",
                 distribution_strategy="auto",
                 requirements_txt="temp.txt",
                 chief_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
@@ -143,9 +143,9 @@ class TestValidate(unittest.TestCase):
     def test_invalid_distribution_strategy(self):
         with pytest.raises(ValueError, match=r"Invalid `distribution_strategy`"):
             validate.validate(
-                entry_point="tests/testdata/mnist_example_using_fit.py",
+                entry_point="python/tensorflow_cloud/tests/testdata/mnist_example_using_fit.py",
                 distribution_strategy="MirroredStrategy",
-                requirements_txt="tests/testdata/requirements.txt",
+                requirements_txt="python/tensorflow_cloud/tests/testdata/requirements.txt",
                 chief_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
                 worker_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
                 worker_count=1,
@@ -159,9 +159,9 @@ class TestValidate(unittest.TestCase):
     def test_invalid_chief_config(self):
         with pytest.raises(ValueError, match=r"Invalid `chief_config`"):
             validate.validate(
-                entry_point="tests/testdata/mnist_example_using_fit.py",
+                entry_point="python/tensorflow_cloud/tests/testdata/mnist_example_using_fit.py",
                 distribution_strategy="auto",
-                requirements_txt="tests/testdata/requirements.txt",
+                requirements_txt="python/tensorflow_cloud/tests/testdata/requirements.txt",
                 chief_config=None,
                 worker_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
                 worker_count=1,
@@ -175,9 +175,9 @@ class TestValidate(unittest.TestCase):
     def test_invalid_worker_config(self):
         with pytest.raises(ValueError, match=r"Invalid `worker_config`"):
             validate.validate(
-                entry_point="tests/testdata/mnist_example_using_fit.py",
+                entry_point="python/tensorflow_cloud/tests/testdata/mnist_example_using_fit.py",
                 distribution_strategy="auto",
-                requirements_txt="tests/testdata/requirements.txt",
+                requirements_txt="python/tensorflow_cloud/tests/testdata/requirements.txt",
                 chief_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
                 worker_config=None,
                 worker_count=1,
@@ -191,9 +191,9 @@ class TestValidate(unittest.TestCase):
     def test_invalid_worker_count(self):
         with pytest.raises(ValueError, match=r"Invalid `worker_count`"):
             validate.validate(
-                entry_point="tests/testdata/mnist_example_using_fit.py",
+                entry_point="python/tensorflow_cloud/tests/testdata/mnist_example_using_fit.py",
                 distribution_strategy="auto",
-                requirements_txt="tests/testdata/requirements.txt",
+                requirements_txt="python/tensorflow_cloud/tests/testdata/requirements.txt",
                 chief_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
                 worker_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
                 worker_count=-1,
@@ -207,9 +207,9 @@ class TestValidate(unittest.TestCase):
     def test_invalid_region(self):
         with pytest.raises(ValueError, match=r"Invalid `region`"):
             validate.validate(
-                entry_point="tests/testdata/mnist_example_using_fit.py",
+                entry_point="python/tensorflow_cloud/tests/testdata/mnist_example_using_fit.py",
                 distribution_strategy="auto",
-                requirements_txt="tests/testdata/requirements.txt",
+                requirements_txt="python/tensorflow_cloud/tests/testdata/requirements.txt",
                 chief_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
                 worker_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
                 worker_count=1,
@@ -223,9 +223,9 @@ class TestValidate(unittest.TestCase):
     def test_invalid_args(self):
         with pytest.raises(ValueError, match=r"Invalid `entry_point_args`"):
             validate.validate(
-                entry_point="tests/testdata/mnist_example_using_fit.py",
+                entry_point="python/tensorflow_cloud/tests/testdata/mnist_example_using_fit.py",
                 distribution_strategy="auto",
-                requirements_txt="tests/testdata/requirements.txt",
+                requirements_txt="python/tensorflow_cloud/tests/testdata/requirements.txt",
                 chief_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
                 worker_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
                 worker_count=1,
@@ -239,9 +239,9 @@ class TestValidate(unittest.TestCase):
     def test_invalid_stream_logs(self):
         with pytest.raises(ValueError, match=r"Invalid `stream_logs`"):
             validate.validate(
-                entry_point="tests/testdata/mnist_example_using_fit.py",
+                entry_point="python/tensorflow_cloud/tests/testdata/mnist_example_using_fit.py",
                 distribution_strategy="auto",
-                requirements_txt="tests/testdata/requirements.txt",
+                requirements_txt="python/tensorflow_cloud/tests/testdata/requirements.txt",
                 chief_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
                 worker_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
                 worker_count=1,
@@ -255,9 +255,9 @@ class TestValidate(unittest.TestCase):
     def test_invalid_cloud_bucket_name(self):
         with pytest.raises(ValueError, match=r"Invalid `docker_image_bucket_name`"):
             validate.validate(
-                entry_point="tests/testdata/mnist_example_using_fit.py",
+                entry_point="python/tensorflow_cloud/tests/testdata/mnist_example_using_fit.py",
                 distribution_strategy="auto",
-                requirements_txt="tests/testdata/requirements.txt",
+                requirements_txt="python/tensorflow_cloud/tests/testdata/requirements.txt",
                 chief_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
                 worker_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
                 worker_count=1,
@@ -271,9 +271,9 @@ class TestValidate(unittest.TestCase):
     def test_invalid_tpu_chief_config(self):
         with pytest.raises(ValueError, match=r"Invalid `chief_config`"):
             validate.validate(
-                entry_point="tests/testdata/mnist_example_using_fit.py",
+                entry_point="python/tensorflow_cloud/tests/testdata/mnist_example_using_fit.py",
                 distribution_strategy="auto",
-                requirements_txt="tests/testdata/requirements.txt",
+                requirements_txt="python/tensorflow_cloud/tests/testdata/requirements.txt",
                 chief_config=machine_config.COMMON_MACHINE_CONFIGS["TPU"],
                 worker_config=machine_config.COMMON_MACHINE_CONFIGS["K80_1X"],
                 worker_count=1,
@@ -287,9 +287,9 @@ class TestValidate(unittest.TestCase):
     def test_invalid_tpu_worker_count(self):
         with pytest.raises(ValueError, match=r"Invalid `worker_count`"):
             validate.validate(
-                entry_point="tests/testdata/mnist_example_using_fit.py",
+                entry_point="python/tensorflow_cloud/tests/testdata/mnist_example_using_fit.py",
                 distribution_strategy="auto",
-                requirements_txt="tests/testdata/requirements.txt",
+                requirements_txt="python/tensorflow_cloud/tests/testdata/requirements.txt",
                 chief_config=machine_config.COMMON_MACHINE_CONFIGS["CPU"],
                 worker_config=machine_config.COMMON_MACHINE_CONFIGS["TPU"],
                 worker_count=2,
@@ -303,9 +303,9 @@ class TestValidate(unittest.TestCase):
     def test_invalid_tpu_accelerator_count(self):
         with pytest.raises(ValueError, match=r"Invalid machine configuration"):
             validate.validate(
-                entry_point="tests/testdata/mnist_example_using_fit.py",
+                entry_point="python/tensorflow_cloud/tests/testdata/mnist_example_using_fit.py",
                 distribution_strategy="auto",
-                requirements_txt="tests/testdata/requirements.txt",
+                requirements_txt="python/tensorflow_cloud/tests/testdata/requirements.txt",
                 chief_config=machine_config.COMMON_MACHINE_CONFIGS["CPU"],
                 worker_config=machine_config.MachineConfig(
                     accelerator_type=machine_config.AcceleratorType.TPU_V3
