+++ python/tensorflow_cloud/tests/testdata/keras_tuner_cifar_example.py
