+++ python/tensorflow_cloud/tests/testdata/mnist_example_using_fit.py
