+++ python/tensorflow_cloud/tests/testdata/mnist_example_using_fit_no_reqs.py
