+++ python/tensorflow_cloud/tests/testdata/save_and_load.py
